Added support for keycode + modifier → keysym lookup.
This involves groveling over the modifier map, storing the current group
modifier, num lock, and caps/shift lock interpretations, and using these
to look up the effective keysym based on the current modifier state.

Note that XKB overrides the interpretation of (some of) these rules, but
that we do not support it at this time. This implementation will therefore
differ from XKB-enabled clients under some circumstances; e.g., we do not
currently support shift-cancels-lock or the use of more than 2 groups.