Completely revamped client property handling.
We now maintain a request (cookie) cache and a value cache in every
ClientWindow instance and use a descriptor class to provide convenient
access to the property values. Requests are made for all of the client
properties during the Withdrawn → Normal state transition, and changes
are tracked via PropertyNotify events.