Cleaned up basic window manager client move/resize/configure tests, and fixed a nasty misinterpretation of ICCCM §4.1.5 that caused a silly bug in configure_notify.
There seems to be a bug with the X server sending us bogus border widths
in the ConfigureNotify generated by a CreateWindow request. This only seems
to be a occur when some client has selected SubstructureRedirect on the
root. To work around this, we set all initial border widths to 1, which
seems to be fine.