Added support for global key and button bindings.
We establish passive grabs for all global bindings on our frames and the
default focus window. One minor but possibly non-trivial side-effect of
this is that we've changed the semantics of per-window event handlers:
if an event is handled by a registered handler, the manager class doesn't
get to see the event at all. This keeps the handlers for grabbed events
from seeing everything else.

Also refactored Bindings class into separate classes for key and button
bindings.