Button bindings do not need the pointer map.
The only thing that actually requires the pointer map is the XTEST
FakeInput request with ButtonPress/ButtonRelease events, which is
used only in the test suite. We now have code in place that does the
translation, but it seems to not be functioning correctly. This only
matters if a non-trivial pointer map is in effect.