We do need SubstructureNotify on frames, after all.
The reason is that without SubstructureNotify, we won't get a
CreateNotify when we create a child of the frame, such as
a title input field window. And without the CreateNotify, it
won't go in the parents tree, and so get_client won't find it,
and so we'll think the client has lost focus when it hasn't.

Sheesh.

This reverts commit 2d54837b1825da4ca78d768cd631e1b0e3e361a1.