Added emergency debugging support.
If we're running outside of a terminal, exploratory and post-mortem
debugging can now be performed via an xterm running in slave mode.
The EmergencyTTY context manager allocates a new pseudo-terminal/terminal
pair, launches an xterm to manage the pty, and attaches our stdin and
stdout to the tty for the duration of the body.

In order to display the new xterm, we need to temporarily disable
SubstructureRedirect on the root window, since we're not responding to
events when we launch it; the new mask_events context manager makes
this simple.

We now also make an effort to cleanly shut down even when an exception
is raised.