Added support for --exec argument to main script.
We now record the command used to launch Dim in the WM_COMMAND
property on the root window. When we want to execute a different
command, we update that property and include a valid timestamp
in the _DIM_WM_EXIT message. The window manager will then call
execvp with the current contents after shutdown.