Added proper default tagset support.
Tag machines now take an additional initarg, default_tagset. It should
be a function of one argument; if tagset lookup fails for some tag T,
the set of values yielded by default_tagset(T) will be used instead.
This replaces the clumsy (and buggy) implicit tagging based on WM_CLASS
with a much more general and flexible mechanism that happens to also
work vastly better for that case.