Ignore EnterNotify in sloppy focus if a LeaveNotify is pending.
The idea here is that if the user just sweeps past a window, it's not
worth focusing it. However, this heuristic is essentially useless on
a fast machine with a decent X server (e.g., my current setup), since
the LeaveNotify doesn't come in soon enough. The real answer here is
some sort of timeout, but I'm not ready to do that yet.