Place windows on the head currently containing the pointer.
We don't want to do this placement for adopted windows on startup,
so the place method now takes an "adopted" flag, along with a
client argument (currently unused, but provided for symmetry with
the other geometry protocol methods). This uncovered the fact that
the adopted flag to the manage method wasn't being properly passed
along in some super calls in a few subclasses.